+++ src/libraries/DecimalMath.sol
@@ -35,23 +35,23 @@ library DecimalMath {
   uint8 public constant highPrecisionDecimals = 27;
 
   /* The number representing 1.0. */
-  uint public constant UNIT = 10 ** uint(decimals);
+  uint256 public constant UNIT = 10 ** uint256(decimals);
 
   /* The number representing 1.0 for higher fidelity numbers. */
-  uint public constant PRECISE_UNIT = 10 ** uint(highPrecisionDecimals);
-  uint private constant UNIT_TO_HIGH_PRECISION_CONVERSION_FACTOR = 10 ** uint(highPrecisionDecimals - decimals);
+  uint256 public constant PRECISE_UNIT = 10 ** uint256(highPrecisionDecimals);
+  uint256 private constant UNIT_TO_HIGH_PRECISION_CONVERSION_FACTOR = 10 ** uint256(highPrecisionDecimals - decimals);
 
   /**
    * @return Provides an interface to UNIT.
    */
-  function unit() external pure returns (uint) {
+  function unit() external pure returns (uint256) {
     return UNIT;
   }
 
   /**
    * @return Provides an interface to PRECISE_UNIT.
    */
-  function preciseUnit() external pure returns (uint) {
+  function preciseUnit() external pure returns (uint256) {
     return PRECISE_UNIT;
   }
 
@@ -64,7 +64,7 @@ library DecimalMath {
    * the internal division always rounds down. This helps save on gas. Rounding
    * is more expensive on gas.
    */
-  function multiplyDecimal(uint x, uint y) internal pure returns (uint) {
+  function multiplyDecimal(uint256 x, uint256 y) internal pure returns (uint256) {
     /* Divide by UNIT to remove the extra factor introduced by the product. */
     return (x * y) / UNIT;
   }
@@ -81,9 +81,9 @@ library DecimalMath {
    * Rounding is useful when you need to retain fidelity for small decimal numbers
    * (eg. small fractions or percentages).
    */
-  function _multiplyDecimalRound(uint x, uint y, uint precisionUnit) private pure returns (uint) {
+  function _multiplyDecimalRound(uint256 x, uint256 y, uint256 precisionUnit) private pure returns (uint256) {
     /* Divide by UNIT to remove the extra factor introduced by the product. */
-    uint quotientTimesTen = (x * y) / (precisionUnit / 10);
+    uint256 quotientTimesTen = (x * y) / (precisionUnit / 10);
 
     if (quotientTimesTen % 10 >= 5) {
       quotientTimesTen += 10;
@@ -104,7 +104,7 @@ library DecimalMath {
    * Rounding is useful when you need to retain fidelity for small decimal numbers
    * (eg. small fractions or percentages).
    */
-  function multiplyDecimalRoundPrecise(uint x, uint y) internal pure returns (uint) {
+  function multiplyDecimalRoundPrecise(uint256 x, uint256 y) internal pure returns (uint256) {
     return _multiplyDecimalRound(x, y, PRECISE_UNIT);
   }
 
@@ -120,7 +120,7 @@ library DecimalMath {
    * Rounding is useful when you need to retain fidelity for small decimal numbers
    * (eg. small fractions or percentages).
    */
-  function multiplyDecimalRound(uint x, uint y) internal pure returns (uint) {
+  function multiplyDecimalRound(uint256 x, uint256 y) internal pure returns (uint256) {
     return _multiplyDecimalRound(x, y, UNIT);
   }
 
@@ -133,7 +133,7 @@ library DecimalMath {
    * this is an integer division, the result is always rounded down.
    * This helps save on gas. Rounding is more expensive on gas.
    */
-  function divideDecimal(uint x, uint y) internal pure returns (uint) {
+  function divideDecimal(uint256 x, uint256 y) internal pure returns (uint256) {
     /* Reintroduce the UNIT factor that will be divided out by y. */
     return (x * UNIT) / y;
   }
@@ -146,8 +146,8 @@ library DecimalMath {
    * is evaluated, so the product of x and the specified precision unit must
    * be less than 2**256. The result is rounded to the nearest increment.
    */
-  function _divideDecimalRound(uint x, uint y, uint precisionUnit) private pure returns (uint) {
-    uint resultTimesTen = (x * (precisionUnit * 10)) / y;
+  function _divideDecimalRound(uint256 x, uint256 y, uint256 precisionUnit) private pure returns (uint256) {
+    uint256 resultTimesTen = (x * (precisionUnit * 10)) / y;
 
     if (resultTimesTen % 10 >= 5) {
       resultTimesTen += 10;
@@ -164,7 +164,7 @@ library DecimalMath {
    * is evaluated, so the product of x and the standard precision unit must
    * be less than 2**256. The result is rounded to the nearest increment.
    */
-  function divideDecimalRound(uint x, uint y) internal pure returns (uint) {
+  function divideDecimalRound(uint256 x, uint256 y) internal pure returns (uint256) {
     return _divideDecimalRound(x, y, UNIT);
   }
 
@@ -176,22 +176,22 @@ library DecimalMath {
    * is evaluated, so the product of x and the high precision unit must
    * be less than 2**256. The result is rounded to the nearest increment.
    */
-  function divideDecimalRoundPrecise(uint x, uint y) internal pure returns (uint) {
+  function divideDecimalRoundPrecise(uint256 x, uint256 y) internal pure returns (uint256) {
     return _divideDecimalRound(x, y, PRECISE_UNIT);
   }
 
   /**
    * @dev Convert a standard decimal representation to a high precision one.
    */
-  function decimalToPreciseDecimal(uint i) internal pure returns (uint) {
+  function decimalToPreciseDecimal(uint256 i) internal pure returns (uint256) {
     return i * UNIT_TO_HIGH_PRECISION_CONVERSION_FACTOR;
   }
 
   /**
    * @dev Convert a high precision decimal to a standard decimal representation.
    */
-  function preciseDecimalToDecimal(uint i) internal pure returns (uint) {
-    uint quotientTimesTen = i / (UNIT_TO_HIGH_PRECISION_CONVERSION_FACTOR / 10);
+  function preciseDecimalToDecimal(uint256 i) internal pure returns (uint256) {
+    uint256 quotientTimesTen = i / (UNIT_TO_HIGH_PRECISION_CONVERSION_FACTOR / 10);
 
     if (quotientTimesTen % 10 >= 5) {
       quotientTimesTen += 10;
